# public/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://storage.fastbuka.com'
paths:
  '/api/v1/storage/{token}':
    get:
      summary: 'Get all storage'
      operationId: getAllStorage
      description: ''
      parameters:
        -
          in: query
          name: env
          description: 'Example api, dev'
          example: dolorem
          required: false
          schema:
            type: string
            description: 'Example api, dev'
            example: dolorem
            nullable: false
      responses:
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Server Error'
                properties:
                  message:
                    type: string
                    example: 'Server Error'
      tags:
        - 'Storage Endpoints'
      security: []
    post:
      summary: 'Store a new storage'
      operationId: storeANewStorage
      description: ''
      parameters:
        -
          in: query
          name: env
          description: 'Example api, dev'
          example: et
          required: false
          schema:
            type: string
            description: 'Example api, dev'
            example: et
            nullable: false
      responses: {  }
      tags:
        - 'Storage Endpoints'
      security: []
    parameters:
      -
        in: path
        name: token
        description: ''
        example: omnis
        required: true
        schema:
          type: string
  '/api/v1/storage/{token}/{storage_uuid}':
    delete:
      summary: 'Delete a storage'
      operationId: deleteAStorage
      description: ''
      parameters:
        -
          in: query
          name: env
          description: 'Example api, dev'
          example: nobis
          required: false
          schema:
            type: string
            description: 'Example api, dev'
            example: nobis
            nullable: false
      responses: {  }
      tags:
        - 'Storage Endpoints'
      security: []
    parameters:
      -
        in: path
        name: token
        description: ''
        example: 8
        required: true
        schema:
          type: integer
      -
        in: path
        name: storage_uuid
        description: ''
        example: 9cbc8977-6f69-35a6-a414-825ec6a30c71
        required: true
        schema:
          type: string
tags:
  -
    name: 'Storage Endpoints'
    description: "\nEndpoints for managing storage"